Refactor naming based on convention + minor bug fix
Refactor names to standard:
ALL CAPS
4 or less letters

Fixed bug related to mouse-wheel (zoom): src\components\DataStructures\common\Renderer\index.js 80

May have other consequences but could not find any
