+++ src/algorithms/explanations/index.js
@@ -1,10 +1,10 @@
 export { default as HSExp } from './HSExp.md';
 export { default as BSTExp } from './BSTExp.md';
 export { default as QSExp } from './QSExp.md';
-export { default as PrimsExp } from './PrimsExp.md';
+export { default as PrimsExp } from './PRIMExp.md';
 export { default as TCExp } from './TCExp.md';
 export { default as BFSSExp } from './BFSSExp.md';
 export { default as QSM3Exp } from './QSM3Exp.md';
 export { default as HSSExp } from './HSSExp.md';
 export { default as DIJKExp } from './DIJKExp.md';
-export {default as ASTARExp} from './ASTARExp.md';
+export {default as ASTARExp} from './ASTExp.md';
