+++ src/algorithms/extra-info/index.js
@@ -1,10 +1,10 @@
 export { default as HSInfo } from './HSInfo.md';
 export { default as BSTInfo } from './BSTInfo.md';
 export { default as QSInfo } from './QSInfo.md';
-export { default as PrimsInfo } from './PrimsInfo.md';
+export { default as PrimsInfo } from './PRIMInfo.md';
 export { default as TCInfo } from './TCInfo.md';
 export { default as BFSSInfo } from './BFSSInfo.md';
 export { default as QSM3Info } from './QSM3Info.md';
 export { default as HSSInfo } from './HSSInfo.md';
 export { default as DIJKInfo } from './DIJKInfo.md';
-export { default as ASTARInfo } from './ASTARInfo.md';
+export { default as ASTARInfo } from './ASTInfo.md';
