+++ src/algorithms/parameters/index.js
@@ -2,10 +2,10 @@
 export { default as HSParam } from './HSParam';
 export { default as BSTParam } from './BSTParam';
 export { default as QSParam } from './QSParam';
-export { default as PrimsParam } from './PrimsParam';
+export { default as PrimsParam } from './PRIMParam';
 export { default as TCParam } from './TCParam';
 export { default as BFSSParam } from './BFSSParam';
 export { default as QSM3Param } from './QSM3Param';
 export { default as HSSParam } from './HSSParam';
 export { default as DIJKParam } from './DIJKParam';
-export { default as ASTARParam } from './ASTARParam';
+export { default as ASTARParam } from './ASTParam';
