+++ src/components/DataStructures/common/Renderer/index.js
@@ -77,7 +77,7 @@ class Renderer extends React.Component {
   }
 
   handleWheel(e) {
-    e.preventDefault();
+    //e.preventDefault();
     const { deltaY } = e;
     this.zoom *= Math.pow(this.zoomFactor, deltaY);
     this.zoom = Math.min(this.zoomMax, Math.max(this.zoomMin, this.zoom));
